+++ rana_qgis_plugin/widgets/settings_dialog.py
@@ -18,10 +18,12 @@ from rana_qgis_plugin.utils_settings import (
     cognito_client_id,
     cognito_client_id_native,
     hcc_working_dir,
+    rana_cache_dir,
     set_base_url,
     set_cognito_client_id,
     set_cognito_client_id_native,
     set_hcc_working_dir,
+    set_rana_cache_dir,
 )
 
 
@@ -62,11 +64,26 @@ class SettingsDialog(QDialog):
         sim_group.layout().addWidget(QLabel("Working directory"), 0, 0)
         self.working_dir_le = QLineEdit(hcc_working_dir(), sim_group)
         sim_group.layout().addWidget(self.working_dir_le, 0, 1)
-        browse_pb = QPushButton("Browse", sim_group)
-        sim_group.layout().addWidget(browse_pb, 0, 2)
-        browse_pb.clicked.connect(self.browse)
+        workdir_browse_pb = QPushButton("Browse", sim_group)
+        sim_group.layout().addWidget(workdir_browse_pb, 0, 2)
+        workdir_browse_pb.clicked.connect(
+            lambda: self.browse("Working Directory", self.working_dir_le)
+        )
 
         layout.addWidget(sim_group)
+
+        files_group = QGroupBox("File storage", self)
+        files_group.setLayout(QGridLayout())
+        files_group.layout().addWidget(QLabel("Cache directory"), 0, 0)
+        self.cache_dir_le = QLineEdit(rana_cache_dir(), files_group)
+        files_group.layout().addWidget(self.cache_dir_le, 0, 1)
+        cachedir_browse_pb = QPushButton("Browse", files_group)
+        files_group.layout().addWidget(cachedir_browse_pb, 0, 2)
+        cachedir_browse_pb.clicked.connect(
+            lambda: self.browse("Cache Directory", self.cache_dir_le)
+        )
+
+        layout.addWidget(files_group)
 
         buttonBox = QDialogButtonBox(
             QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel
@@ -94,19 +111,20 @@ class SettingsDialog(QDialog):
             set_base_url(self.url_lineedit.text())
 
         set_hcc_working_dir(self.working_dir_le.text())
+        set_rana_cache_dir(self.cache_dir_le.text())
 
         return super().accept()
 
-    def browse(self):
-        work_dir = QFileDialog.getExistingDirectory(
-            self, "Select Working Directory", self.working_dir_le.text()
+    def browse(self, dir_human_name: str, dir_box_le: QLineEdit):
+        dir = QFileDialog.getExistingDirectory(
+            self, f"Select {dir_human_name}", dir_box_le.text()
         )
-        if work_dir:
-            if not is_writable(work_dir):
+        if dir:
+            if not is_writable(dir):
                 QMessageBox.warning(
                     self,
                     "Warning",
                     "Can't write to the selected location. Please select a folder to which you have write permission.",
                 )
                 return
-            self.working_dir_le.setText(work_dir)
+            dir_box_le.setText(dir)