+++ HISTORY.rst
@@ -13,6 +13,7 @@ History
 - Add support for selecting and uploading multiple files to Rana
 - Add option to create a new folder
 - Add button and dialog from M&S plugin to generate/upload new schematisation to 3Di and Rana
+- Allow user to set file cache directory
 
 
 1.1.20 (2025-11-18)